Renamed contrib apps
