+++ wagtail/contrib/frontendcache/apps.py
@@ -1,9 +1,9 @@
 from django.apps import AppConfig
-from wagtail.contrib.wagtailfrontendcache.signal_handlers import register_signal_handlers
+from wagtail.contrib.frontendcache.signal_handlers import register_signal_handlers
 
 
 class WagtailFrontendCacheAppConfig(AppConfig):
-    name = 'wagtail.contrib.wagtailfrontendcache'
+    name = 'wagtail.contrib.frontendcache'
     label = 'wagtailfrontendcache'
     verbose_name = "Wagtail frontend cache"
