+++ wagtail/contrib/frontendcache/signal_handlers.py
@@ -1,0 +1,24 @@
+from django.db import models
+
+from wagtail.wagtailcore.signals import page_published, page_unpublished
+
+from wagtail.contrib.frontendcache.utils import purge_page_from_cache
+
+
+def page_published_signal_handler(instance, **kwargs):
+    purge_page_from_cache(instance)
+
+
+def page_unpublished_signal_handler(instance, **kwargs):
+    purge_page_from_cache(instance)
+
+
+def register_signal_handlers():
+    # Get list of models that are page types
+    Page = models.get_model('wagtailcore', 'Page')
+    indexed_models = [model for model in models.get_models() if issubclass(model, Page)]
+
+    # Loop through list and register signal handlers for each one
+    for model in indexed_models:
+        page_published.connect(page_published_signal_handler, sender=model)
+        page_unpublished.connect(page_unpublished_signal_handler, sender=model)
