+++ wagtail/contrib/frontendcache/tests.py
@@ -1,8 +1,8 @@
 from django.test import TestCase
 from django.test.utils import override_settings
 
-from wagtail.contrib.wagtailfrontendcache.utils import get_backends
-from wagtail.contrib.wagtailfrontendcache.backends import HTTPBackend, CloudflareBackend
+from wagtail.contrib.frontendcache.utils import get_backends
+from wagtail.contrib.frontendcache.backends import HTTPBackend, CloudflareBackend
 
 
 class TestBackendConfiguration(TestCase):
