+++ wagtail/contrib/routablepage/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 
 class WagtailRoutablePageAppConfig(AppConfig):
-    name = 'wagtail.contrib.wagtailroutablepage'
+    name = 'wagtail.contrib.routablepage'
     label = 'wagtailroutablepage'
     verbose_name = "Wagtail routablepage"
