+++ wagtail/contrib/sitemaps/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 
 class WagtailSitemapsAppConfig(AppConfig):
-    name = 'wagtail.contrib.wagtailsitemaps'
+    name = 'wagtail.contrib.sitemaps'
     label = 'wagtailsitemaps'
     verbose_name = "Wagtail sitemaps"
