+++ wagtail/contrib/staticsitegen/__init__.py
