+++ wagtail/contrib/staticsitegen/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailStaticSiteGenAppConfig(AppConfig):
+    name = 'wagtail.contrib.staticsitegen'
+    label = 'wagtailstaticsitegen'
+    verbose_name = "Wagtail static site generator"
