+++ wagtail/contrib/styleguide/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 
 class WagtailStyleGuideAppConfig(AppConfig):
-    name = 'wagtail.contrib.wagtailstyleguide'
+    name = 'wagtail.contrib.styleguide'
     label = 'wagtailstyleguide'
     verbose_name = "Wagtail style guide"
