+++ wagtail/contrib/styleguide/locale/en/LC_MESSAGES/django.po
