+++ wagtail/contrib/wagtailstyleguide/wagtail_hooks.py
@@ -1,20 +1,1 @@
-from django.conf.urls import url
-from django.core import urlresolvers
-from django.utils.translation import ugettext_lazy as _
-
-from wagtail.wagtailcore import hooks
-from wagtail.wagtailadmin.menu import MenuItem
-
-from . import views
-
-
-@hooks.register('register_admin_urls')
-def register_admin_urls():
-    return [
-        url(r'^styleguide/$', views.index, name='wagtailstyleguide'),
-    ]
-
-
-@hooks.register('register_admin_menu_item')
-def register_styleguide_menu_item():
-    return MenuItem(_('Styleguide'), urlresolvers.reverse('wagtailstyleguide'), classnames='icon icon-image', order=1000)
+from wagtail.contrib.styleguide.wagtail_hooks import *
